separate out more pieces
