+++ browser/resources/settings/email_aliases_page/email_aliases.tsx
@@ -4,17 +4,9 @@
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import { createRoot } from 'react-dom/client';
-import { getLocale } from '$web-common/locale'
-import { MainEmailEntryForm } from './content/email_aliases_signin_page'
-import { spacing, font, radius, typography }
-  from '@brave/leo/tokens/css/variables'
+import { ManagePage } from './content/email_aliases_manage_page'
+import { StyleSheetManager } from 'styled-components'
 import * as React from 'react'
-import Card from './content/styles/Card'
-import Col from './content/styles/Col'
-import { MainView } from './content/email_aliases_main_view'
-
-import SecureLink from '$web-common/SecureLink'
-import styled, { StyleSheetManager } from 'styled-components'
 import {
   AuthenticationStatus,
   AuthState,
@@ -25,42 +17,7 @@ import {
   EmailAliasesServiceObserverInterface
 } from "gen/brave/components/email_aliases/email_aliases.mojom.m"
 
-const PageCol = styled(Col)`
-  font: ${font.default.regular};
-  padding: 0;
-  margin: 0;
-  & h4 {
-    font: ${font.heading.h4};
-    line-height: ${typography.heading.h4.lineHeight};
-    margin: 0;
-  }
-`
-
-const SectionTitle = styled(Card)`
-  border-radius: ${radius.m};
-  padding: ${spacing['2Xl']} ${spacing['2Xl']} ${spacing.l} ${spacing['2Xl']};
-  display: flex;
-  flex-direction: column;
-  row-gap: ${spacing.m};
-`
-
-const Introduction = () =>
-  <SectionTitle>
-      <div>
-      <h4>{getLocale('emailAliasesShortDescription')}</h4>
-      </div>
-      <div>
-        {getLocale('emailAliasesDescription')}  {
-           /* TODO(https://github.com/brave/brave-browser/issues/45408):
-           // Link to the email aliases support page */}
-        <SecureLink href="https://support.brave.com" target='_blank'>
-          {getLocale('emailAliasesLearnMore')}
-        </SecureLink>
-      </div>
-  </SectionTitle>
-
-export const ManagePage = ({ emailAliasesService, bindObserver }:
-  {
+export const ManagePageConnected = ({ emailAliasesService, bindObserver }: {
     emailAliasesService: EmailAliasesServiceInterface,
     bindObserver: (observer: EmailAliasesServiceObserverInterface) => () => void
   }) => {
@@ -79,18 +36,10 @@ export const ManagePage = ({ emailAliasesService, bindObserver }:
     return bindObserver(observer)
   }, [] /* Only run at mount. */)
   return (
-    <PageCol>
-      <Introduction />
-      {authState.status === AuthenticationStatus.kUnauthenticated ||
-       authState.status === AuthenticationStatus.kAuthenticating
-        ? <MainEmailEntryForm
-            authState={authState}
-            emailAliasesService={emailAliasesService} />
-        : <MainView
-            authState={authState}
-            aliasesState={aliasesState}
-            emailAliasesService={emailAliasesService} />}
-    </PageCol>
+    <ManagePage
+      authState={authState}
+      aliasesState={aliasesState}
+      emailAliasesService={emailAliasesService} />
   )
 }
 
@@ -104,8 +53,9 @@ export const mount = (at: HTMLElement) => {
   }
   root.render(
     <StyleSheetManager target={at}>
-      <ManagePage emailAliasesService={emailAliasesService}
-                  bindObserver={bindObserver} />
+      <ManagePageConnected
+        emailAliasesService={emailAliasesService}
+        bindObserver={bindObserver} />
     </StyleSheetManager>
   )
 }
